Merging the 2D and 3D NODE training procedures. More below.
* `node_baseline_3d.yaml`
Updating the 3D configuration. Removing the `initial_conditions`,
`description`, `experiment_name` and `reconstruct` configurations
since they are not used in this case. Also, changed the number of
training steps to a more reasonable value (20k)

* `node-warp-train.py`
- Removing unused package imports.

- Added early stopping option and recording it on the configuration
saved in the results folder.

- Checking the input data dimensionality (using the network
configuration actually) to avoid instatiating the `WarpingDataset`
instance, as it is only needed for blending the 2D cases. Also
avoiding reading the reconstruction configuration (removed from the
3D configuration case, see above).

- Simplifying the correspondences test: if they don't exist, we simply
raise an exception. Use the other scripts to generate them, this way
the training script only trains the warping, simplifying our lives.

- Saving the weights with best loss results so far as well.