Added information on MegaDepth dataset download. More below.
* `README.md`
Added information on how to download and process the MegaDepth test
set and process it for our trainings. Did the same for the FRLL
dataset and started the NeRSemble section as well, althought that is
still in progress.

* `Makefile`
Fixed the FRLL cropping rule to refer to the `crop-face-images.py`
script instead of the `align.py` script.

* `standalone/megadepth_generation/crop.sh`
Added sanity check for when no cmdline parameter is provided.